15-utilities/Mamey.Image.BackgroundRemoval/requirements.txt
rembg>=2.0.50
onnxruntime>=1.16.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
click>=8.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
